JIRA-89: Improve error handling when we have Rights issue
* Move JIRABadRequestException and JIRAConnectionException as an API
