+++ jira-macro/jira-macro-count/src/main/java/org/xwiki/contrib/jira/count/internal/JIRACountMacro.java
@@ -31,8 +31,8 @@ import org.xwiki.component.annotation.Component;
 import org.xwiki.contrib.jira.config.JIRAServer;
 import org.xwiki.contrib.jira.count.JIRACountMacroParameters;
 import org.xwiki.contrib.jira.macro.internal.HTTPJIRAFetcher;
-import org.xwiki.contrib.jira.macro.internal.JIRABadRequestException;
-import org.xwiki.contrib.jira.macro.internal.JIRAConnectionException;
+import org.xwiki.contrib.jira.macro.JIRABadRequestException;
+import org.xwiki.contrib.jira.macro.JIRAConnectionException;
 import org.xwiki.contrib.jira.macro.internal.JIRAErrorGenerator;
 import org.xwiki.contrib.jira.macro.internal.JIRAMacroTransformationManager;
 import org.xwiki.contrib.jira.macro.internal.JIRAURLHelper;
