+++ jira-macro/jira-macro-default/src/main/java/org/xwiki/contrib/jira/macro/internal/source/AbstractJIRADataSource.java
@@ -31,8 +31,8 @@ import org.xwiki.contrib.jira.macro.JIRADataSource;
 import org.xwiki.contrib.jira.macro.JIRAField;
 import org.xwiki.contrib.jira.macro.JIRAMacroParameters;
 import org.xwiki.contrib.jira.macro.internal.HTTPJIRAFetcher;
-import org.xwiki.contrib.jira.macro.internal.JIRABadRequestException;
-import org.xwiki.contrib.jira.macro.internal.JIRAConnectionException;
+import org.xwiki.contrib.jira.macro.JIRABadRequestException;
+import org.xwiki.contrib.jira.macro.JIRAConnectionException;
 import org.xwiki.contrib.jira.macro.internal.JIRAURLHelper;
 import org.xwiki.rendering.macro.MacroExecutionException;
